add image proc

+++ multi_hand_tracking.py
@@ -6,6 +6,7 @@ import shutil
 from pprint import pprint
 from scripts import convertProtobufToJson
 
+
 def multi_hand_track(input_video_path, input_basename,  output_dir):
     command = " ".join([
         f'/usr/local/bazel/2.0.0/lib/bazel/bin/bazel build -c opt',
@@ -15,7 +16,7 @@ def multi_hand_track(input_video_path, input_basename,  output_dir):
     res = subprocess.run(command, stderr=subprocess.STDOUT, shell=True)
 
     if os.path.exists(f"{output_dir}/result"):
-        shutil.rmtree(f"{output_dir}/result")
+        shutil.rmtree(f"{output_dir}/result"    )
 
     command_options = [
         f'./bazel-bin/mediapipe/examples/desktop/multi_hand_tracking/multi_hand_tracking_cpu',
@@ -33,7 +34,8 @@ def multi_hand_track(input_video_path, input_basename,  output_dir):
 def doMultiHandTracking(input_video_path):
     input_dir = os.path.dirname(input_video_path)
     input_basename = os.path.basename(input_video_path)
-    output_dir = input_video_path.replace(".mp4", "-mp4")
+    output_dir = input_video_path[:input_video_path.rfind(
+        ".")] + "-" + input_video_path[input_video_path.rfind(".")+1:]
 
     if not os.path.exists(output_dir):
         os.makedirs(output_dir)
@@ -43,10 +45,11 @@ def doMultiHandTracking(input_video_path):
     print("convert protobuf files")
     convertProtobufToJson.convertFilesInDir(f"{output_dir}/result/")
 
-    shutil.copy2(f"{output_dir}/{input_basename}", f"{output_dir}/video.mp4")
-
-    os.makedirs(f"{input_dir}/handtracked/", exist_ok=True)
-    shutil.copy2(f"{output_dir}/{input_basename}", f"{input_dir}/handtracked/{input_basename}")
+    if os.path.exists(f"{output_dir}/{input_basename}"):
+        shutil.copy2(f"{output_dir}/{input_basename}", f"{output_dir}/video.mp4")
+        os.makedirs(f"{input_dir}/handtracked/", exist_ok=True)
+        shutil.copy2(f"{output_dir}/{input_basename}",
+                    f"{input_dir}/handtracked/{input_basename}")
 
 
 if __name__ == "__main__":
@@ -59,4 +62,3 @@ if __name__ == "__main__":
     print(input_video_path)
 
     doMultiHandTracking(input_video_path)
-
